refactor/bolt-agent
