+++ neo4j/internal/bolt/bolt5_test.go
@@ -23,7 +23,6 @@ import (
 	"context"
 	"fmt"
 	iauth "github.com/neo4j/neo4j-go-driver/v5/neo4j/internal/auth"
-	"github.com/neo4j/neo4j-go-driver/v5/neo4j/internal/boltagent"
 	idb "github.com/neo4j/neo4j-go-driver/v5/neo4j/internal/db"
 	"github.com/neo4j/neo4j-go-driver/v5/neo4j/notifications"
 	"io"
@@ -178,13 +177,11 @@ func TestBolt5(outer *testing.T) {
 			handshake := srv.waitForHandshake()
 			AssertVersionInHandshake(t, handshake, 5, 3)
 			srv.acceptVersion(5, 3)
-			// 5.3+ hello must contain mandatory bolt_agent dictionary and mandatory product field
+			// 5.3 hello must contain mandatory bolt_agent dictionary and mandatory product field
 			hmap := srv.waitForHelloWithoutAuthToken()
-			boltAgent, exists := hmap["bolt_agent"].(map[string]any)
-			if !exists {
-				panic("Missing bolt_agent in hello")
-			}
-			AssertStringEqual(t, boltAgent["product"].(string), boltagent.Product)
+			boltAgent, exists := hmap["bolt_agent"]
+			AssertTrue(t, exists)
+			AssertStringContain(t, boltAgent.(map[string]any)["product"].(string), "neo4j-go/")
 			srv.acceptHello()
 
 			srv.waitForLogon()
