+++ neo4j/internal/boltagent/bolt_agent_test.go
@@ -8,13 +8,14 @@ import (
 func init() {
 	os = "darwin"
 	arch = "amd64"
-	version = "go1.20.3"
+	goVersion = "go1.20.3"
+	driverVersion = "5.9.0"
 }
 
 func TestNew(t *testing.T) {
 	actual := New()
 
-	AssertStringEqual(t, actual.Product(), Product)
+	AssertStringEqual(t, actual.Product(), "neo4j-go/5.9.0")
 	AssertStringEqual(t, actual.Platform(), "darwin; amd64")
 	AssertStringEqual(t, actual.Language(), "Go/go1.20.3")
 }
