+++ neo4j/internal/metadata/metadata.go
@@ -1,3 +1,3 @@
-package internal
+package metadata
 
 const DriverVersion = "5.9.0"
